+++ CHANGELOG.md
@@ -5,3 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.2.2](https://github.com/utnet-org/utility-sdk-rs/compare/unc-sdk-v2.2.1...unc-sdk-v2.2.2) - 2024-06-12
+
+### Other
+- Update deprecated versions in code
+- Update deprecated versions in code